% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to a database}
\usage{
connect_to_db(..., .envir = parent.frame())
}
\arguments{
\item{...}{arguments passed on to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{.envir}{passed on to \code{\link[withr:defer]{withr::defer()}}}
}
\value{
A \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}} object
}
\description{
General helper to connect to a database through \code{\link[DBI:dbConnect]{DBI::dbConnect()}}, while ensuring
that the connection is closed when the connection object goes out of scope.
}
