% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillIn.R
\name{fillIn}
\alias{fillIn}
\title{fillIn}
\usage{
fillIn(UV)
}
\arguments{
\item{datetime}{POSIXct date/time}
}
\description{
Fill in any missing rows so that dataset is a complete 5-min dataset
}
\examples{
\dontrun{library(dataRetrieval)
site <- '040851385'
parameterCd <- '00060'
UVPList <- c("00010","00045","00060","00065","00095","00300","00400","63680","99133")
startDate <- "2015-09-01"
endDate <- "2015-09-10"
UV <- readNWISuv(site, UVPList, startDate="", endDate="")
UV_2 <- removeDups(UV)
}

}
