% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{generate_concepts}
\alias{generate_concepts}
\title{Generate the \code{omopcat_concepts} table}
\usage{
generate_concepts(cdm, concept_ids)
}
\arguments{
\item{cdm}{A \code{\link{CDMConnector}} object, e.g. from \code{\link[CDMConnector:cdm_from_con]{CDMConnector::cdm_from_con()}}}

\item{concept_ids}{A vector of concept IDs}
}
\value{
A \code{data.frame} with the monthly counts
}
\description{
Generate the \code{omopcat_concepts} table
}
\keyword{internal}
