% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{generate_summary_stats}
\alias{generate_summary_stats}
\title{Generate the \code{omopcat_summary_stats} table}
\usage{
generate_summary_stats(cdm)
}
\arguments{
\item{cdm}{A \code{\link{CDMConnector}} object, e.g. from \code{\link[CDMConnector:cdm_from_con]{CDMConnector::cdm_from_con()}}}
}
\value{
A \code{data.frame} with the summary statistics
}
\description{
Generate the \code{omopcat_summary_stats} table
}
\keyword{internal}
