% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omopcat.preprocessing-package.R
\docType{package}
\name{omopcat.preprocessing-package}
\alias{omopcat.preprocessing}
\alias{omopcat.preprocessing-package}
\title{omopcat.preprocessing: Preprocessing for the omopcat App}
\description{
Preprocess data for the omopcat app.
}
\author{
\strong{Maintainer}: Milan Malfait \email{m.malfait@ucl.ac.uk}

Authors:
\itemize{
  \item Stefan Piatek
  \item Baptiste Briot Ribeyre
  \item Andy South
}

Other contributors:
\itemize{
  \item SAFEHR [copyright holder, funder]
}

}
\keyword{internal}
