% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDups.R
\name{removeDups}
\alias{removeDups}
\title{removeDups}
\usage{
removeDups(UV)
}
\arguments{
\item{datetime}{POSIXct date/time}
}
\description{
Remove duplicate date/times
}
\examples{
\dontrun{library(dataRetrieval)
site <- '040851385'
parameterCd <- '00060'
UVPList <- c("00010","00045","00060","00065","00095","00300","00400","63680","99133")
startDate <- "2015-09-01"
endDate <- "2015-09-10"
UV <- readNWISuv(site, UVPList, startDate="", endDate="")
UV_2 <- removeDups(UV)
}

}
