% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-low_frequencies.R
\name{replace_low_frequencies}
\alias{replace_low_frequencies}
\title{Helper function to replace low frequencies}
\usage{
replace_low_frequencies(df, cols, threshold, replacement)
}
\arguments{
\item{df}{A data.frame}

\item{cols}{A character vector of column names for which to apply the replacement}

\item{threshold}{Threshold value below which values will be replaced by \code{replacement}}

\item{replacement}{Value with which values below \code{threshold} will be replaced}
}
\value{
A numeric vector with low frequencies replaced
}
\description{
Replaces frequencies below the given \code{threshold} with the value given by the \code{replacement}
argument. These values are typically set by the \code{LOW_FREQUENCY_THRESHOLD}
and \code{LOW_FREQUENCY_REPLACEMENT} environment variables. This is done to avoid identifiability
of the health records.
Values equal to zero are filtered out prior to replacement.
}
\keyword{internal}
