% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_counts.R
\name{summarise_counts}
\alias{summarise_counts}
\title{Summarise record counts}
\usage{
summarise_counts(omop_table, concept_col, date_col, level)
}
\arguments{
\item{omop_table}{A table from the OMOP CDM}

\item{concept_col}{The name of the concept column to calculate statistics for}

\item{date_col}{The name of the date column to calculate statistics for}

\item{level}{The resolution at which to summarise the record counts.
Currently supports \code{"monthly"} or \code{"quarterly"}}
}
\value{
A \code{data.frame} with the following columns:
\itemize{
\item \code{concept_id}: The concept ID
\item \code{concept_name}: The concept name
\item \code{date_year}: The year of the date
\item \code{date_month} or \code{date_quarter}: The month or quarter of the date, depending on \code{level}
\item \code{person_count}: The number of unique patients per concept for each month
\item \code{records_per_person}: The average number of records per person per concept for each month
}
}
\description{
Summarise record counts
}
\keyword{internal}
